backend/requirements.txt
asgiref==3.5.0
Django>=3.2.0,<3.3.0
djangorestframework>=3.12.0,<3.13.0
django-filter>=2.4.0,<3.0.0
pytz==2022.1
sqlparse==0.4.2
ipython==8.2.0
djangorestframework-simplejwt==5.1.0
psycopg2-binary
cryptography==3.4.8
gunicorn==20.1.0
gunicorn[gevent]
django-cors-headers
kavenegar
python-memcached
